+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `VersionConfig.due_date_for(risk)` to calculate until when the finding with the given risk should be remediated.
+- Add *graphics* pandoc filter to the default settings.
+
+### Changed
+
+- Update documentation.
 
 ## [0.2.7] - 2025-05-26